refactor(pipeline): decompose build into collect/preprocess/render/index
Extract the monolithic build_sync_impl_sync into named stages so later
mdBook-parity increments target one stage each. Zero behaviour change:
output is byte-identical to the pre-A baseline (72/72 assets).

- src/pipeline: collect, run_sync, async index, identity preprocess seam
- src/render: markdown (mdast+syntect) and html (Tera/assets) modules
- src/core: Args + build orchestration only
- Tests relocated with their code; preprocess identity covered

Refs #1
