+++ .agent/handoffs/2026-08-08-mdbook-parity-a.md
@@ -1,0 +1,42 @@
+# Handover: mdBook parity Increment A
+
+**Date**: 2026-08-08  
+**UTC end**: see verification report  
+**Change slug**: mdbook-parity-a  
+**Issue**: #1  
+**Branch**: `task/1-pipeline-decomposition`
+
+## Completed
+
+- Decomposed build into `src/pipeline/` + `src/render/`
+- Identity preprocess seam with tests
+- Byte-identical output vs pre-A baseline (72 files)
+- Verification, validation, review artefacts written
+- Unrelated WIP left in `git stash` (message: `wip: unrelated SEO/template/dist changes`)
+
+## Known-good state
+
+- All unit + integration tests listed in verification report pass
+- Clippy clean with `-D warnings`
+
+## Resume steps (next agent)
+
+1. `git checkout task/1-pipeline-decomposition` (or merge PR for #1)
+2. Claim Gitea **#2** (Increment B — SUMMARY.md book model); Phase 2.5 findings already in the plan
+3. Do **not** drop `stash@{0}` without restoring for the user — it holds SEO/template work
+4. B1 starts with structure fixture `tests/fixtures/test_book_mdbook.structure.json`
+
+## Artefacts
+
+- Research: `docs/plans/mdbook-parity-research.md`
+- Design: `docs/plans/mdbook-parity-implementation-plan.md`
+- Session: `.agent/sessions/2026-08-08-mdbook-parity-a.md`
+- Verification: `docs/verification/verification-report-mdbook-parity-a.md`
+- Validation: `docs/validation/validation-report-mdbook-parity-a.md`
+- Review: `docs/plans/review-mdbook-parity-a.md`
+
+## Next actions
+
+1. Open/merge PR for #1  
+2. Start Increment B on `task/2-summary-book-model`  
+3. Restore stash only if resuming SEO WIP separately
